# business/catalogue-service-micro/src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Catalogue Service API

# Describe your paths here
paths:
...
  /catalogue/{partyId}/default:
    get:
      operationId: getDefaultCatalogue
      summary: Get default catalogue for party
      description: Retrieves the default catalgoue for the specified party. The catalague is supposed to have and ID field with "default" value and be compliant with UBL standard.
      produces:
      - application/json
      parameters:
      - in: path
        name: partyId
        description: Identifier of the party
        required: true
        type: string
      responses:
        200:
          description: Default the default catalogue for the specified party
        204:
          description: No default catalogue exists for the specified party
  /catalogue/{standard}:
    post:
      operationId: addXMLCatalogue
      summary: Add catalogue
      description: Adds the catalogue represented in XML format. The serialized catalogue should be compliant with the specified standard
      consumes: [application/json, application/xml]
      produces: [application/json, application/xml]
      parameters:
      - in: path
        name: standard
        required: true
      - in: body
        name: serializedCatalogue
        required: true
      responses:
        201:
          description: Saved catalogue is returned
        400:
          description: Invalid content type header or invalid standard
    put:
      operationId: updateJSONCatalogue
      summary: Update JSON catalogue
      description: Updates the catalogue represented in JSON serialization. The serialization should be compliant with the default standard, which is UBL
      consumes: [application/json]
      produces: [application/json]
      parameters:
      - in: body
        name: catalogueJson
        required: true
      responses:
        304:
          description: Updated catalogue is returned
        400:
          description: Invalid standard or invalid catalogue serialization
        501:
          description: An unsupported standard is provided
  /catalogue/{standard}/{uuid}:
    get:
      operationId: getCatalogue
      summary: Returns the catalogue with the given standard and uuid
      description: Retrieves the catalogue for the given standard and uuid.
      produces:
      - application/json
      parameters:
      - in: path
        name: standard
        description: Standard of the requested catalogue
        required: true
        type: string
      - in: path
        name: uuid
        description: Identifier of the catalogue
        required: true
        type: string
      responses:
        200:
          description: The catalogue specified with the uuid is returned
        204:
          description: No catalogue exists for the given uuid
        400:
          description: Invalid standard
    delete:
      operationId: deleteCatalogue
      summary: Delete catalogue
      description: Deletes the catalogue for the given standard and uuid.
      parameters:
      - in: path
        name: standard
        description: Standard of the catalogue to be deleted
        required: true
        type: string
      - in: path
        name: uuid
        description: Identifier of the catalogue to be deleted
        required: true
        type: string
      responses:
        200:
          description: Catalogue deleted, if one exists
        400:
          description: Invalid standard
  /catalogue/template:
    get:
      operationId: downloadTemplate
      summary: Generate template
      description: Generates an excel-based template for the specified categories. Category ids and taxonomy ids must be provided in comma separated manner and they must have the same number of elements. Taxonomy id must be provided such that they specify the taxonomies including the specified categories.
      produces:
      - application/octet-stream
      parameters:
      - in: query
        name: categoryIds
        description: Comma separated category ids. Example "http://www.aidimme.es/FurnitureSectorOntology.owl#MDFBoard,0173-1#01-ACH237#011"
        required: false
        type: string
      - in: query
        name: taxonomyIds
        description: Comma separated taxonomy ids. Example "FurnitureOntology,eClass"
        required: false
        type: string
      responses:
        200:
          description: Returned the generated template
  /catalogue/template/upload:
    post:
      operationId: uploadTemplate
      summary: Upload template
      description: Adds the catalogue specified with the provided template. The created catalogue is compliant with the default standard, which is UBL
      consumes: [multipart/form-data]
      parameters:
      - in: formData
        name: file
        description: Filled in excel-based template
        required: true
        type: file
      - in: query
        name: partyId
        description: Identifier of the party submitting the template
        required: false
        type: string
      - in: query
        name: partyName
        description: Name of the party submitting the template
        required: false
        type: string
      responses:
        201:
          description: Returned the generated template
...